# lap4/contracts/team-gate.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus — Access Gate Service API
  version: 1.0.0
  description: |
    Access Gate Service cung cấp API cho Core Business để tra cứu log ra/vào,
    trạng thái cổng, thông tin thẻ, và quản lý đăng ký thẻ mới trong
    Smart Campus Operations Platform.
    Service nhận sự kiện quẹt thẻ, kiểm tra hợp lệ, ghi log, và cung cấp
    dữ liệu audit theo chuẩn Problem Details khi có lỗi.
  contact:
    name: Access Gate API Team
    email: access-gate@campus.local
servers:
  - url: https://api.campus.local
    description: Production API endpoint
  - url: http://localhost:4011
    description: Local mock server
tags:
  - name: health
    description: Kiểm tra và giám sát trạng thái dịch vụ
  - name: access-logs
    description: Truy vấn log ra/vào
  - name: gates
    description: Truy vấn trạng thái cổng
  - name: cards
    description: Truy vấn và quản lý thông tin thẻ
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Kiểm tra trạng thái service
      description: Endpoint để kiểm tra service còn sống.
      tags: [health]
      security: []
      responses:
        '200':
          description: Service hoạt động bình thường
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
              examples:
                ok:
                  value:
                    status: ok
                    service: access-gate-service
                    time: '2026-08-11T11:37:00Z'

  /access/logs/recent:
    get:
      operationId: listRecentAccessLogs
      summary: Lấy danh sách log ra/vào gần nhất
      description: |
        Core Business gọi endpoint này để audit log ra/vào theo thời gian thực.
        Hỗ trợ cursor-based pagination.
      tags: [access-logs]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/GateIdFilter'
      responses:
        '200':
          description: Danh sách log ra/vào
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessLogPage'
              examples:
                recentLogs:
                  value:
                    items:
                      - logId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                        cardId: RFID-2026-001
                        gateId: GATE-01
                        direction: IN
                        status: GRANTED
                        reasonCode: VALID_CARD
                        operatorNote: null
                        timestamp: '2026-08-11T07:30:00Z'
                    nextCursor: null
                    hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /access/logs/{logId}:
    get:
      operationId: getAccessLogById
      summary: Lấy chi tiết một log ra/vào
      description: Tra cứu chi tiết 1 log cụ thể theo logId để audit sâu.
      tags: [access-logs]
      parameters:
        - $ref: '#/components/parameters/LogId'
      responses:
        '200':
          description: Chi tiết log ra/vào
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessLog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /gates/{gateId}/status:
    get:
      operationId: getGateStatus
      summary: Lấy trạng thái hiện tại của một cổng
      description: Core Business gọi endpoint này để biết cổng đang mở/đóng/lỗi.
      tags: [gates]
      parameters:
        - $ref: '#/components/parameters/GateId'
      responses:
        '200':
          description: Trạng thái cổng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /cards/{cardId}:
    get:
      operationId: getCardById
      summary: Lấy thông tin một thẻ
      description: Core Business gọi endpoint này để xác minh thông tin chủ thẻ.
      tags: [cards]
      parameters:
        - $ref: '#/components/parameters/CardId'
      responses:
        '200':
          description: Thông tin thẻ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /cards:
    post:
      operationId: registerCard
      summary: Đăng ký thẻ mới
      description: |
        Admin đăng ký thẻ mới vào hệ thống Access Gate.
        Hỗ trợ 2 loại thẻ: Student và Staff, phân biệt qua discriminator cardType.
      tags: [cards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCardRequest'
            examples:
              studentCard:
                value:
                  cardType: STUDENT
                  cardId: RFID-2026-001
                  personId: SV001
                  studentCode: DNU2026001
              staffCard:
                value:
                  cardType: STAFF
                  cardId: RFID-2026-002
                  personId: NV045
                  department: FIT
      responses:
        '201':
          description: Thẻ đã được đăng ký
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    LogId:
      name: logId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID của log ra/vào.
    GateId:
      name: gateId
      in: path
      required: true
      schema:
        type: string
        pattern: '^GATE-[0-9]{2}$'
      description: Mã cổng.
    CardId:
      name: cardId
      in: path
      required: true
      schema:
        type: string
        pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
      description: Mã thẻ RFID.
    GateIdFilter:
      name: gateId
      in: query
      required: false
      schema:
        type: string
        pattern: '^GATE-[0-9]{2}$'
      description: Lọc log theo mã cổng.
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: [string, 'null']
        minLength: 1
        maxLength: 200
      description: Cursor phiên bản tiếp theo cho pagination.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Số bản ghi tối đa trả về mỗi trang.

  schemas:
    HealthStatus:
      type: object
      additionalProperties: false
      required: [status, service, time]
      properties:
        status:
          type: string
          enum: [ok]
          example: ok
        service:
          type: string
          example: access-gate-service
        time:
          type: string
          format: date-time
          example: '2026-08-11T11:37:00Z'

    AccessLog:
      type: object
      additionalProperties: false
      required: [logId, cardId, gateId, direction, status, reasonCode, timestamp]
      properties:
        logId:
          type: string
          format: uuid
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
        direction:
          type: string
          enum: [IN, OUT]
        status:
          type: string
          enum: [GRANTED, DENIED]
        reasonCode:
          type: string
          enum:
            - VALID_CARD
            - EXPIRED_CARD
            - BLACKLISTED
            - OUTSIDE_TIME_WINDOW
            - UNKNOWN_CARD
            - POLICY_DENY
        operatorNote:
          type: [string, 'null']
        timestamp:
          type: string
          format: date-time

    AccessLogPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccessLog'
        nextCursor:
          type: [string, 'null']
        hasMore:
          type: boolean

    GateStatus:
      type: object
      additionalProperties: false
      required: [gateId, state, updatedAt]
      properties:
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
        state:
          type: string
          enum: [OPEN, CLOSED, FAULT]
        lastCardId:
          type: [string, 'null']
        updatedAt:
          type: string
          format: date-time

    Card:
      type: object
      additionalProperties: false
      required: [cardId, personId, cardType, status]
      properties:
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
        personId:
          type: string
        cardType:
          type: string
          enum: [STUDENT, STAFF, VISITOR]
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        issuedAt:
          type: [string, 'null']
          format: date-time
        expiresAt:
          type: [string, 'null']
          format: date-time

    RegisterCardRequest:
      oneOf:
        - $ref: '#/components/schemas/StudentCardRequest'
        - $ref: '#/components/schemas/StaffCardRequest'
      discriminator:
        propertyName: cardType
        mapping:
          STUDENT: '#/components/schemas/StudentCardRequest'
          STAFF: '#/components/schemas/StaffCardRequest'

    StudentCardRequest:
      type: object
      additionalProperties: false
      required: [cardType, cardId, personId, studentCode]
      properties:
        cardType:
          type: string
          const: STUDENT
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
        personId:
          type: string
        studentCode:
          type: string

    StaffCardRequest:
      type: object
      additionalProperties: false
      required: [cardType, cardId, personId, department]
      properties:
        cardType:
          type: string
          const: STAFF
        cardId:
          type: string
          pattern: '^RFID-[0-9]{4}-[0-9]{3}$'
        personId:
          type: string
        department:
          type: string

    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: [string, 'null']
        instance:
          type: [string, 'null']
          format: uri
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemField'

    ProblemField:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string

  responses:
    BadRequest:
      description: Dữ liệu sai định dạng
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/validation
            title: Dữ liệu không hợp lệ
            status: 400
            detail: Payload không đúng JSON Schema
            instance: https://api.campus.local/cards
            errors:
              - field: cardId
                code: PATTERN_MISMATCH
                message: cardId phải có dạng RFID-YYYY-NNN

    Unauthorized:
      description: Thiếu hoặc sai thông tin xác thực
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/unauthorized
            title: Chưa xác thực
            status: 401
            detail: Thiếu Bearer token
            instance: https://api.campus.local/access/logs/recent
            errors: []

    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/not-found
            title: Không tìm thấy tài nguyên
            status: 404
            detail: cardId không tồn tại trong hệ thống
            instance: https://api.campus.local/cards/RFID-2026-999
            errors: []

    Conflict:
      description: Thẻ đã tồn tại trong hệ thống
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/conflict
            title: Thẻ đã tồn tại
            status: 409
            detail: cardId đã được đăng ký trước đó
            instance: https://api.campus.local/cards
            errors: []

    UnprocessableEntity:
      description: Payload đúng định dạng nhưng vi phạm nghiệp vụ
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/business-rule
            title: Vi phạm nghiệp vụ
            status: 422
            detail: studentCode không tồn tại trong hệ thống quản lý sinh viên
            instance: https://api.campus.local/cards
            errors:
              - field: studentCode
                code: NOT_FOUND_IN_SIS
                message: Mã sinh viên không hợp lệ

    InternalServerError:
      description: Lỗi không mong muốn phía server
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/internal
            title: Lỗi hệ thống
            status: 500
            detail: Đã xảy ra lỗi không mong muốn
            instance: https://api.campus.local/access/logs/recent
            errors: []